# consensus and crypto perf optimizations