--- a/src/Catalyst.Abstractions/Cryptography/ICryptoContext.cs
+++ b/src/Catalyst.Abstractions/Cryptography/ICryptoContext.cs
@@ -21,6 +21,8 @@
 
 #endregion
 
+using System;
+
 namespace Catalyst.Abstractions.Cryptography
 {
     public interface ICryptoContext
@@ -101,21 +103,8 @@ namespace Catalyst.Abstractions.Cryptography
         /// <param name="message"></param>
         /// <param name="context"></param>
         /// <returns></returns>
-        ISignature Sign(IPrivateKey privateKey, byte[] message, byte[] context);
+        ISignature Sign(IPrivateKey privateKey, ReadOnlySpan<byte> message, ReadOnlySpan<byte> context);
 
-        /// <summary>
-        ///     Signs message using provided private key and returns the signature.
-        /// </summary>
-        /// <param name="privateKey"></param>
-        /// <param name="message"></param>
-        /// <param name="messageLength"></param>
-        /// <param name="context"></param>
-        /// <param name="contextLength"></param>
-        /// <returns></returns>
-        ISignature Sign(IPrivateKey privateKey, byte[] message, int messageLength, byte[] context, int contextLength);
-
-        bool Verify(ISignature signature, byte[] message, byte[] context);
-        
-        bool Verify(ISignature signature, byte[] message, int messageLength, byte[] context, int contextLength);
+        bool Verify(ISignature signature, ReadOnlySpan<byte> message, ReadOnlySpan<byte> context);
     }
 }