--- a/src/Catalyst.Abstractions/Hashing/IHashProvider.cs
+++ b/src/Catalyst.Abstractions/Hashing/IHashProvider.cs
@@ -37,5 +37,6 @@ namespace Catalyst.Abstractions.Hashing
         MultiHash ComputeMultiHash(IEnumerable<byte> content);
         MultiHash Cast(byte[] data);
         bool IsValidHash(byte[] data);
+        MultiHash ComputeMultiHash(byte[] data, int offset, int count);
     }
 }