--- a/src/Catalyst.Benchmark/Program.cs
+++ b/src/Catalyst.Benchmark/Program.cs
@@ -26,7 +26,7 @@ using BenchmarkDotNet.Running;
 
 namespace Catalyst.Benchmark
 {
-    public class Program
+    public static class Program
     {
         public static void Main(string[] args)
         {