… src/Catalyst.Benchmark/Catalyst.Benchmark.csproj
… src/Catalyst.Benchmark/Catalyst.Core.Modules.Hashing/HashProviderBenchmark.cs
… src/Catalyst.Core.Lib.Tests/Fakes/FakeKeySigner.cs
… src/Catalyst.Core.Lib.Tests/IntegrationTests/P2P/IO/Transport/Channels/PeerClientChannelFactoryTests.cs
… src/Catalyst.Core.Lib.Tests/IntegrationTests/P2P/IO/Transport/Channels/PeerServerChannelFactoryTests.cs
… src/Catalyst.Core.Lib.Tests/IntegrationTests/P2P/PeerValidationIntegrationTest.cs
… src/Catalyst.Core.Lib.Tests/IntegrationTests/Rpc/IO/Transport/Channels/RpcClientChannelFactoryTests.cs
… src/Catalyst.Core.Lib.Tests/IntegrationTests/Rpc/IO/Transport/Channels/RpcServerChannelFactoryTests.cs
… src/Catalyst.Core.Lib.Tests/UnitTests/IO/Handlers/ProtocolMessageSignHandlerTests.cs
… src/Catalyst.Core.Lib.Tests/UnitTests/IO/Handlers/ProtocolMessageVerifyHandlerTests.cs
… src/Catalyst.Core.Lib.Tests/UnitTests/P2P/IO/Messaging/Broadcast/BroadcastHandlerTests.cs
… src/Catalyst.Core.Lib.Tests/UnitTests/P2P/IO/Messaging/Broadcast/BroadcastManagerTests.cs
… src/Catalyst.Core.Lib.Tests/UnitTests/P2P/IO/Transport/Channels/PeerClientChannelFactoryTests.cs
… src/Catalyst.Core.Lib.Tests/UnitTests/P2P/IO/Transport/Channels/PeerServerChannelFactoryTests.cs
… src/Catalyst.Core.Lib.Tests/UnitTests/Rpc/IO/Observers/SignMessageRequestObserverTests.cs
… src/Catalyst.Core.Lib.Tests/UnitTests/Rpc/IO/Observers/VerifyMessageRequestObserverTests.cs
… src/Catalyst.Core.Lib.Tests/UnitTests/Rpc/IO/Transport/Channels/RpcClientChannelFactoryTests.cs
… src/Catalyst.Core.Lib.Tests/UnitTests/Rpc/IO/Transport/Channels/RpcServerChannelFactoryTests.cs
… src/Catalyst.Core.Lib.Tests/UnitTests/Validators/TransactionValidatorTests.cs
… src/Catalyst.Core.Lib/Extensions/Protocol/Wire/ProtocolMessageExtensions.cs
… src/Catalyst.Core.Lib/Extensions/Protocol/Wire/TransactionBroadcastExtensions.cs
… src/Catalyst.Core.Lib/IO/Handlers/ProtocolMessageVerifyHandler.cs
… src/Catalyst.Core.Lib/Validators/TransactionValidator.cs
… src/Catalyst.Core.Modules.Consensus/Deltas/DeltaBuilder.cs
… src/Catalyst.Core.Modules.Consensus/Deltas/DeltaCache.cs
… src/Catalyst.Core.Modules.Cryptography.BulletProofs.Tests/UnitTests/CryptoWrapperTests.cs
… src/Catalyst.Core.Modules.Cryptography.BulletProofs/Catalyst.Core.Modules.Cryptography.BulletProofs.csproj
… src/Catalyst.Core.Modules.Cryptography.BulletProofs/FfiWrapper.cs
… src/Catalyst.Core.Modules.Cryptography.BulletProofs/NativeBinding.cs
… src/Catalyst.Core.Modules.Hashing.Tests/UnitTests/HashingTests.cs