### New Pull Request Submissions:

1. [x] Have you followed the guidelines in our [Contributing document](https://github.com/atlascity/Community/tree/master/CONTRIBUTING.md)?
2. [x] Have you checked to ensure there aren't other open [Pull Requests](../../pulls) for the same update/change?
3. [x] I have added tests to cover my changes.
4. [x] All new and existing tests passed.
5. [x] Have you lint your code locally prior to submission?
6. [x] Does your code follows the code style of this project?
7. [ ] Does your change require a change to the documentation.
    - [ ] I have updated the documentation accordingly.
9. [x] Have you added an explanation of what your changes do and why you'd like us to include them?
10. [ ] Have you inserted a keyword and link to the issues the PR closes in its descriptions (ex `closes #1`) ?
11. [x] Is you branch up to date, have you integrated all the latest changes from develop and resolved conflicts ?

* **What kind of change does this PR introduce?** (Bug fix, feature, docs update, ...)

This PR changes methods of `IHashProvider` and `ICryptoContext` that enable them to hash / sigh / verify messages without allocating memory via `.ToByteArray()`. 

Additionally, for the hasher, one method is added that enables a concatenation of a message byte payload with a suffix. This method can be used to greatly improved (benchmark added) an execution of `DeltaBuilder`.

* **What is the current behavior?** (You can also link to an open issue here)

Currently, in several places there are calls `.ToByteArray`, to pass a proto message as `byte[]` to a method. 

* **What is the new behavior (if this is a feature change)?**

This PR changes several cases, to use `ReadOnlySpan<byte>` that allows:

1. using a message serialized an array from to `ArrayPool<byte>.Shared`
1. using `ByteString.Span` without cloning its content
1. using stack-allocated span whenever possible (currently, not used)

* **Does this PR introduce a breaking change?** (What changes might users need to make in their application due to this PR?)

The public API that is changed is following:

1. `IKeySigner`
1. `IHashProvider` 
1. `ICryptoContext`

The hash provider has not Span based overloads, as underlying Blake implementation based on Bouncy Castle, does not currently use currently [TryComputeHash](https://docs.microsoft.com/en-us/dotnet/api/system.security.cryptography.hashalgorithm.trycomputehash).

* **Other information**:

There are a few more methods that could be possibly spanified in similar manner, like: `ICryptoContext.GetSignatureFromBytes(byte[] signatureBytes, byte[] publicKeyBytes)`

Some benchmarks that were added, that might be useful for future improvement. In the second case, the embedded serialization

### Benchmark: calculating a hash for a single item + salt in DeltaBuilder

|          Method |       Mean |    Error |   StdDev |  Gen 0 |  Allocated |
|---------------- |-----------:|---------:|---------:|-------:|--------:|
|   Concat_manual | 1,798.5 ns | 35.33 ns | 55.00 ns | 0.2956 |   928 B |
| Concat_embedded |   309.4 ns |  7.65 ns | 21.72 ns | 0.0587 |      184 B |

### Benchmark: Sign/Verify chain 

|                                 Method |     Mean |    Error |   StdDev |   Median | Allocated |
|--------------------------------------- |---------:|---------:|---------:|---------:|----------:|
|            SignVerify_with_ToByteArray | 704.2 ns | 14.01 ns | 37.87 ns | 685.9 ns |     624 B |
| SignVerify_with_embedded_serialization | 744.2 ns | 14.61 ns | 28.49 ns | 743.1 ns |      96 B |