@CLAassistant 2019-11-26
[![CLA assistant check](https://cla-assistant.io/pull/badge/not_signed)](https://cla-assistant.io/catalyst-network/Catalyst.Framework?pullRequest=1111) <br/>Thank you for your submission, we really appreciate it. Like many open source projects, we ask that you sign our [Contributor License Agreement](https://cla-assistant.io/catalyst-network/Catalyst.Framework?pullRequest=1111) before we can accept your contribution.<br/><sub>You have signed the CLA already but the status is still pending? Let us [recheck](https://cla-assistant.io/check/catalyst-network/Catalyst.Framework?pullRequest=1111) it.</sub>