@Scooletz 2019-11-26
@Xela101 This somewhat overlaps with https://github.com/catalyst-network/Catalyst.Framework/pull/1111 as it changes a bit the way salted message hash is calculated, but it's a minor thing and I can easily address it once your or mine PR is merged.