@nshCore 2019-11-26
@Scooletz thanks for the PR I am going to check it out locally and run, first glances looks good!