@Scooletz 2019-11-26
I'll clear this tomorrow. Maybe I'll come with a better abstraction for the call sites, something that does not use `IMessage`.