@Scooletz 2019-11-27
Let me try to add spanning. 1-2 hours most