@Scooletz 2019-11-27
I pushed the spanified version. I see that PR Quality Review is failing, but will amend it tomorrow. I updated the PR description as well to match the new Span-ified API.