@nshCore 2019-11-28
![Codacy](https://app.codacy.com/assets/images/favicon.png) Here is an overview of what got changed by this pull request:


```diff

Issues
======
+ Solved 2
- Added 34
           

Complexity increasing per file
==============================
- src/Catalyst.Abstractions/KeySigner/KeySignerExtensions.cs  1
- src/Catalyst.Abstractions/Hashing/HashExtensions.cs  1
- src/Catalyst.Benchmark/Catalyst.Abstractions/HashingBenchmark.cs  1
- src/Catalyst.Abstractions/Cryptography/CryptoExtensions.cs  1
- src/Catalyst.Core.Modules.Cryptography.BulletProofs/NativeBinding.cs  2
- src/Catalyst.Benchmark/Program.cs  1
- src/Catalyst.Protocol/Extensions.cs  1
- src/Catalyst.Core.Lib.Tests/Fakes/FakeKeySigner.cs  2
- src/Catalyst.Benchmark/Catalyst.Abstractions/CryptoBenchmark.cs  1
         

Clones added
============
- src/Catalyst.Core.Lib.Tests/UnitTests/IO/Handlers/ProtocolMessageVerifyHandlerTests.cs  2
- src/Catalyst.Core.Lib.Tests/UnitTests/Rpc/IO/Observers/SignMessageRequestObserverTests.cs  3
- src/Catalyst.Core.Lib.Tests/UnitTests/Validators/TransactionValidatorTests.cs  3
- src/Catalyst.Core.Lib.Tests/UnitTests/Rpc/IO/Transport/Channels/RpcClientChannelFactoryTests.cs  1
- src/Catalyst.Core.Lib.Tests/UnitTests/P2P/IO/Transport/Channels/PeerServerChannelFactoryTests.cs  2
- src/Catalyst.Abstractions/Cryptography/CryptoExtensions.cs  2
- src/Catalyst.Core.Lib.Tests/UnitTests/IO/Handlers/ProtocolMessageSignHandlerTests.cs  2
- src/Catalyst.Core.Modules.KeySigner.Tests/UnitTests/KeySignerTests.cs  3
- src/Catalyst.Core.Lib.Tests/IntegrationTests/P2P/PeerValidationIntegrationTest.cs  1
- src/Catalyst.Core.Lib.Tests/UnitTests/Rpc/IO/Transport/Channels/RpcServerChannelFactoryTests.cs  1
- src/Catalyst.Core.Lib.Tests/UnitTests/P2P/IO/Messaging/Broadcast/BroadcastHandlerTests.cs  2
- src/Catalyst.Benchmark/Catalyst.Abstractions/CryptoBenchmark.cs  2
         

Clones removed
==============
+ src/Catalyst.Core.Lib.Tests/IntegrationTests/P2P/IO/Transport/Channels/PeerServerChannelFactoryTests.cs  -4
+ src/Catalyst.Core.Lib.Tests/IntegrationTests/P2P/IO/Transport/Channels/PeerClientChannelFactoryTests.cs  -1
+ src/Catalyst.Core.Lib.Tests/UnitTests/P2P/IO/Transport/Channels/PeerClientChannelFactoryTests.cs  -2
+ src/Catalyst.Core.Lib.Tests/IntegrationTests/Rpc/IO/Transport/Channels/RpcClientChannelFactoryTests.cs  -2
+ src/Catalyst.Core.Lib.Tests/IntegrationTests/Rpc/IO/Transport/Channels/RpcServerChannelFactoryTests.cs  -5
         
```


See the complete overview on [Codacy](https://app.codacy.com/app/Catalyst.Framework/Catalyst.Framework/pullRequest?prid=4569935&bid=15361671)